allow collision checking before computing grasp quality to be toggled
There are cases where the grasp quality must be computed even when there is an apparent collision (e.g. imprecise reconstruction of a 3D scene). This commit adds two services, getCheckCollision and setCheckCollision, that allows the controlling node to get and toggle whether collision check is to be performed prior to computing the grasp quality metrics. To be in line with current conventions, collision checking is enabled by default.